Nodrop framework, SCP113
Credit to LorenLuke for the nodrop framework, glad to finally find some use for your work.

Adds SCP 113, using the current nodrop framework. There's a bit of work to be done in terms of bugs, and maybe adding some organ damage later as is consistent with documentation.

http://www.scp-wiki.net/scp-113

Know issues: Nodrop doesn't account for trying to put an object on a table, or in a container. I'll look into changing that.